Added a function for finding isosurface intersects
Right now it only gives the fraction of the distance from
a high to low vertex value at which the isovalue is.  (Linearly
interpolated.)  This doesn't take into account the actual positions
of the vertices in space.

The positions of the vertices in space will have to be exploited when
actually computing the volume of the intersecting region.